MBX-2530 Move ABTest model
