+++ MindboxTests/InAppMessages/Tests/InAppConfigResponseTests/InAppConfigResponseTests.swift
@@ -59,27 +59,27 @@ class InAppConfigResponseTests: XCTestCase {
                                                                                      intentPayload: "3",
                                                                                      type: "simpleImage")]))]
         
-        let abTestObject1 = InAppConfigResponse.ABTest.ABTestVariant.ABTestObject(
+        let abTestObject1 = ABTest.ABTestVariant.ABTestObject(
             type: .inapps,
             kind: .all,
             inapps: ["inapp1", "inapp2"]
         )
-
+        
         // Создаем структуры ABTestVariant
-        let abTestVariant1 = InAppConfigResponse.ABTest.ABTestVariant(
-            id: "1", modulus: InAppConfigResponse.ABTest.ABTestVariant.Modulus(lower: 0, upper: 50),
+        let abTestVariant1 = ABTest.ABTestVariant(
+            id: "1", modulus: ABTest.ABTestVariant.Modulus(lower: 0, upper: 50),
             objects: [abTestObject1]
         )
-
-        let abTestVariant2 = InAppConfigResponse.ABTest.ABTestVariant(
-            id: "2", modulus: InAppConfigResponse.ABTest.ABTestVariant.Modulus(lower: 50, upper: 100),
+        
+        let abTestVariant2 = ABTest.ABTestVariant(
+            id: "2", modulus: ABTest.ABTestVariant.Modulus(lower: 50, upper: 100),
             objects: [abTestObject1]
         )
-        let abtests: [InAppConfigResponse.ABTest]? = [.init(id: "id123",
-                                                            sdkVersion: .init(min: 1, max: nil),
-                                                            salt: "salt123",
-                                                            variants: [abTestVariant1,
-                                                                       abTestVariant2]),
+        let abtests: [ABTest]? = [.init(id: "id123",
+                                        sdkVersion: .init(min: 1, max: nil),
+                                        salt: "salt123",
+                                        variants: [abTestVariant1,
+                                                   abTestVariant2]),
         ]
         
         let monitoring = Monitoring(logs: [.init(requestId: "request1",
@@ -92,8 +92,8 @@ class InAppConfigResponseTests: XCTestCase {
                                                  to: "destination2")])
         
         let settings = Settings(operations: .init(viewProduct: .init(systemName: "product"),
-                                                                      viewCategory: .init(systemName: "category"),
-                                                                      setCart: .init(systemName: "cart")))
+                                                  viewCategory: .init(systemName: "category"),
+                                                  setCart: .init(systemName: "cart")))
         
         XCTAssertEqual(response.inapps, inapps)
         XCTAssertEqual(response.abtests, abtests)
@@ -104,13 +104,13 @@ class InAppConfigResponseTests: XCTestCase {
     func test_config_settings_invalid_to_parse() throws {
         let response = try getConfig(name: "InappConfigResponseSettingsInvalid")
         let inapps: [InApp]? = [.init(id: "6f93e2ef-0615-4e63-9c80-24bcb9e83b83",
-                                                         sdkVersion: SdkVersion(min: 4, max: nil),
-                                                         targeting: .and(AndTargeting(nodes: [.true(TrueTargeting())])),
-                                                         form: InApp.InAppFormVariants(variants: [.init(imageUrl: "1",
-                                                                                                                      redirectUrl: "2",
-                                                                                                                      intentPayload: "3",
-                                                                                                                      type: "simpleImage")]))]
-
+                                      sdkVersion: SdkVersion(min: 4, max: nil),
+                                      targeting: .and(AndTargeting(nodes: [.true(TrueTargeting())])),
+                                      form: InApp.InAppFormVariants(variants: [.init(imageUrl: "1",
+                                                                                     redirectUrl: "2",
+                                                                                     intentPayload: "3",
+                                                                                     type: "simpleImage")]))]
+        
         XCTAssertEqual(response.inapps, inapps)
         // No systemName in Settings JSON
         XCTAssertNil(response.settings)
@@ -119,13 +119,13 @@ class InAppConfigResponseTests: XCTestCase {
     func test_config_monitoring_invalid_to_parse() throws {
         let response = try getConfig(name: "InappConfigResponseMonitoringInvalid")
         let inapps: [InApp]? = [.init(id: "6f93e2ef-0615-4e63-9c80-24bcb9e83b83",
-                                                         sdkVersion: SdkVersion(min: 4, max: nil),
-                                                         targeting: .and(AndTargeting(nodes: [.true(TrueTargeting())])),
-                                                         form: InApp.InAppFormVariants(variants: [.init(imageUrl: "1",
-                                                                                                                      redirectUrl: "2",
-                                                                                                                      intentPayload: "3",
-                                                                                                                      type: "simpleImage")]))]
-
+                                      sdkVersion: SdkVersion(min: 4, max: nil),
+                                      targeting: .and(AndTargeting(nodes: [.true(TrueTargeting())])),
+                                      form: InApp.InAppFormVariants(variants: [.init(imageUrl: "1",
+                                                                                     redirectUrl: "2",
+                                                                                     intentPayload: "3",
+                                                                                     type: "simpleImage")]))]
+        
         XCTAssertEqual(response.inapps, inapps)
         // No id in Monitoring JSON
         XCTAssertNil(response.monitoring)
@@ -134,13 +134,13 @@ class InAppConfigResponseTests: XCTestCase {
     func test_config_abtests_invalid_to_parse() throws {
         let response = try getConfig(name: "InappConfigResponseAbtestsInvalid")
         let inapps: [InApp]? = [.init(id: "6f93e2ef-0615-4e63-9c80-24bcb9e83b83",
-                                                         sdkVersion: SdkVersion(min: 4, max: nil),
-                                                         targeting: .and(AndTargeting(nodes: [.true(TrueTargeting())])),
-                                                         form: InApp.InAppFormVariants(variants: [.init(imageUrl: "1",
-                                                                                                                      redirectUrl: "2",
-                                                                                                                      intentPayload: "3",
-                                                                                                                      type: "simpleImage")]))]
-
+                                      sdkVersion: SdkVersion(min: 4, max: nil),
+                                      targeting: .and(AndTargeting(nodes: [.true(TrueTargeting())])),
+                                      form: InApp.InAppFormVariants(variants: [.init(imageUrl: "1",
+                                                                                     redirectUrl: "2",
+                                                                                     intentPayload: "3",
+                                                                                     type: "simpleImage")]))]
+        
         XCTAssertEqual(response.inapps, inapps)
         // No id in Abtest JSON
         XCTAssertNil(response.abtests)
